Armar el query detallado para consulta de ejecución presupuestal

Armar el query detallado para consulta de ejecución presupuestal

make_gasto_query(
  years = NULL,
  nivel = NULL,
  sector = NULL,
  pliego = NULL,
  unidad_ejecutora = NULL,
  goblocal_o_manc = NULL,
  mancomunidad = NULL,
  departamento = NULL,
  provincia = NULL,
  municipalidad = NULL,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL,
  fuente_financiamiento = NULL,
  rubro = NULL,
  tipo_de_recurso = NULL,
  generica = NULL,
  subgenerica = NULL,
  detalle_subgenerica = NULL,
  especifica = NULL,
  detalle_especifica = NULL,
  departamento_meta = NULL,
  trimestre = NULL,
  mes = NULL
)

Arguments

nivel

chr. Nivel de gobierno. Puede ser "E" (gobierno nacional), "M" (gobiernos locales) o "R" (gobiernos regionales). Si es "E" o "R", le corresponde sector. Si es "M", le corresponde goblocal_o_manc.

sector

chr. Codigo de sector, de forma "XX" (2).

pliego

chr. Código de pliego, de forma "XXX" (3).

unidad_ejecutora

chr. Código de Unidad Ejecutora. No tiene forma estandar

goblocal_o_manc

chr. Escoger entre "M" (gobierno local) o "N" (mancomunidad). Si es "M", desbloque departamento, provincia y municipalidad. Si es "N", le corresponde mancomunidad.

mancomunidad

chr. Código de mancomunidad, de forma "XXXXXX" (6).

departamento

chr. Código de departamento. El valor "07" corresponde a Callao.

provincia

chr. Código de provincia. El valor "01" corresponde a capital de departamento.

municipalidad

chr. Código de municipalidad, de forma "XXXXXX". No corresponde con ubigeo.

categoria_presupuestal

chr. Código de programa presupuestal de forma "XXXX" (4). Para Acciones Centrales usar "9001" y para APnoP usar "9002".

producto

chr. Código de producto o proyecto de inversión de forma "XXXXXXX" (7).

actividad

chr. Código de actividad, acción de inversión u obra, de forma "XXXXXXX" (7).

funcion

chr. Código de función de forma "XX" (2).

division_funcional

chr. Código de división funcional de forma "XXX" (3).

grupo_funcional

chr. Código de grupo funcional de forma "XXXX" (4).

meta

chr. Código de meta. No tiene forma estandarizada.

fuente_financiamiento

int. Código de fuente de financiamiento. En su versión más reciente va del 1 al 5.

rubro

chr. código de rubro, de forma "XX" (2).

tipo_de_recurso

chr, Código de tipo de recurso. No tiene forma estandar, puede ser de uno o dos caracteres, que pueden ser letras o números.

generica

chr. Código de genérica de forma "X-X-X"

subgenerica

chr. Código de sub-genérica de forma "X-X-X"

detalle_subgenerica

chr. Código de detalle de sub-genérica de forma "X-X-X-X"

especifica

chr. Código de específica de forma "X-X-X-X-X"

detalle_especifica

chr. Código de detalle de específica de forma "X-X-X-X-X-X"

departamento_meta

chr. Código de departamento. Para obtener "Callao" usar "07", y para obtener 'Exterior' usar "98".

trimestre

int. Número del 1 al 4

mes

int. Número del 1 al 12

year

Año del cual se quiere obtener la información

Value

Una lista cuyos elementos tienen el key-value de los parámetros de la API