Elegir periodo anual de consulta

Elegir periodo anual de consulta

Usage

elegir_periodo_anual(x, periodo = NULL)

Arguments

x

Objeto de clase <transpaeco> recipiente de los parámetros de la consulta deseada

periodo

numeric. Periodo anual de consulta. Actualmente solo se puede consultar a partir del 2012.

Value

Un data.frame con clase <transpaeco>

Examples

iniciar_transparencia_economica() %>%
    elegir_periodo_anual(periodo = 2022)
#> 
#> ── Seguimiento al gasto presupuestal (actualizacion diaria) ────────────────────
#> 
#> ── Parametros de consulta ──
#> 
#> • Periodo anual: 2022
#> 
#> ── Data ──
#> 
#> ℹ No se ha ejecutado ninguna consulta

iniciar_transparencia_economica() %>%
    elegir_periodo_anual(periodo = 2022) %>%
    consultar()
#> ℹ Iniciando consulta
#> ℹ Unificando consultas...
#> ✔ Consultas realizadas y unificadas
#> # A tibble: 1 × 10
#>   periodo total          pia          pim certificacion compromiso_anual
#>     <dbl> <chr>        <dbl>        <dbl>         <dbl>            <dbl>
#> 1    2022 TOTAL 197002269014 237822055719  225523647131     215896701775
#> # ℹ 4 more variables: atencion_de_compromiso_mensual <dbl>, devengado <dbl>,
#> #   girado <dbl>, avance_percent <dbl>

# pueden consultarse multiples periodos al mismo tiempo
# por ejemplo, el periodo de pandemia por Covid 19
iniciar_transparencia_economica() %>%
    elegir_periodo_anual(periodo = 2020:2022) %>%
    consultar()
#> ℹ Iniciando consulta
#> ⠙ 2/3 ETA:  1s | Ejecutando consulta  
#> ℹ Unificando consultas...
#> ✔ Consultas realizadas y unificadas
#> # A tibble: 3 × 10
#>   periodo total          pia          pim certificacion compromiso_anual
#>     <int> <chr>        <dbl>        <dbl>         <dbl>            <dbl>
#> 1    2020 TOTAL 177367859707 217254208912  199474122166     191603364023
#> 2    2021 TOTAL 183029770158 227932217391  212472939298     205036051241
#> 3    2022 TOTAL 197002269014 237822055719  225523647131     215896701775
#> # ℹ 4 more variables: atencion_de_compromiso_mensual <dbl>, devengado <dbl>,
#> #   girado <dbl>, avance_percent <dbl>

# se emite mensaje de error si se intenta elegir periodos no abarcados
tryCatch({
    iniciar_transparencia_economica(modulo = "gasto") %>%
        elegir_periodo_anual(periodo = 2011)
}, error = function(e) print(e))
#> <error/rlang_error>
#> Error in `elegir_periodo_anual()`:
#> ! Periodo anual debe estar entre 2012 y 2023
#> ---
#> Backtrace:
#>      ▆
#>   1. ├─pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
#>   2. │ └─pkgdown::build_site(...)
#>   3. │   └─pkgdown:::build_site_local(...)
#>   4. │     └─pkgdown::build_reference(...)
#>   5. │       └─purrr::map(...)
#>   6. │         └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>   7. │           ├─purrr:::with_indexed_errors(...)
#>   8. │           │ └─base::withCallingHandlers(...)
#>   9. │           ├─purrr:::call_with_cleanup(...)
#>  10. │           └─pkgdown (local) .f(.x[[i]], ...)
#>  11. │             ├─base::withCallingHandlers(...)
#>  12. │             └─pkgdown:::data_reference_topic(...)
#>  13. │               └─pkgdown:::run_examples(...)
#>  14. │                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  15. │                   └─downlit::evaluate_and_highlight(...)
#>  16. │                     └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  17. │                       └─evaluate:::evaluate_call(...)
#>  18. │                         ├─evaluate (local) timing_fn(...)
#>  19. │                         ├─evaluate (local) handle(...)
#>  20. │                         │ └─base::try(f, silent = TRUE)
#>  21. │                         │   └─base::tryCatch(...)
#>  22. │                         │     └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  23. │                         │       └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  24. │                         │         └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  25. │                         ├─base::withCallingHandlers(...)
#>  26. │                         ├─base::withVisible(...)
#>  27. │                         └─evaluate:::eval_with_user_handlers(expr, envir, enclos, user_handlers)
#>  28. │                           └─base::eval(expr, envir, enclos)
#>  29. │                             └─base::eval(expr, envir, enclos)
#>  30. │                               ├─base::tryCatch(...)
#>  31. │                               │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  32. │                               │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  33. │                               │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  34. │                               └─iniciar_transparencia_economica(modulo = "gasto") %>% ...
#>  35. └─perutranspaeconomica::elegir_periodo_anual(., periodo = 2011)